move and rename functions for transforming CollectionVenues
